Use python api instead of xml for much faster (> 4x) result parsing.
The main downside is lessened security if running python 2.5 or earlier:
because the result data is eval()'d, arbitrary code could be executed if
an attacker could either control the hudson server or could inject data
into the response. On python 2.6 and later we use ast.literal_eval which
does not have this vulnerability.